OrderServiceImplTest completion

+++ pneuservis-persistence/src/main/java/cz/fi/muni/pa165/pneuservis/dao/OrderDAOImpl.java
@@ -39,6 +39,7 @@ public class OrderDAOImpl implements OrderDAO {
         if (order == null) {
             throw new IllegalArgumentException("Cannot delete null order");
         }
+        if (order.getId() == null) throw new IllegalArgumentException("Cannot delete order with null id.");
         em.remove(order);
     }
